Ran clippy and format

+++ src/star_field.rs
@@ -24,13 +24,12 @@ fn setup_star_field(
     let num_x_tiles: i32 = (resolution.screen_dimensions.x / WIDTH).floor() as i32 + 1;
     let num_y_tiles: i32 = (resolution.screen_dimensions.y / HEIGHT).floor() as i32 + 1;
 
-
     for x in -num_x_tiles..num_x_tiles {
         for y in -num_y_tiles..num_x_tiles {
-            let position = Vec3::new(x as f32 * WIDTH,y as f32 * HEIGHT, ZINDEX);
-            let rotation_radians: f32 = ((x + y) as f32 % 4 as f32 * 90.0).to_radians();
+            let position = Vec3::new(x as f32 * WIDTH, y as f32 * HEIGHT, ZINDEX);
+            let rotation_radians: f32 = ((x + y) as f32 % 4.0 * 90.0).to_radians();
             let mut transform: Transform = Transform::from_translation(position)
-                    .with_scale(Vec3::splat(resolution.pixel_ratio));
+                .with_scale(Vec3::splat(resolution.pixel_ratio));
             transform.rotate(Quat::from_rotation_z(rotation_radians));
 
             commands.spawn((
